+++ src/app/cart/cart.module.ts
@@ -1,14 +1,14 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { CartComponent, CartListComponent, CartItemComponent } from './components';
+import { CartComponent, CartListComponent, CartProductComponent } from './components';
 
 import { SharedModule } from '../shared/shared.module';
 
 const declarations = [
   CartComponent,
   CartListComponent,
-  CartItemComponent,
+  CartProductComponent,
 ];
 
 const imports = [