+++ src/app/cart/components/cart-list/index.ts
@@ -1,1 +1,0 @@
-export { CartListComponent } from './cart-list.component';