+++ src/app/cart/components/index.ts
@@ -1,3 +1,3 @@
-export * from './cart';
-export * from './cart-list';
-export * from './cart-item';
+export * from './cart/cart.component';
+export * from './cart-list/cart-list.component';
+export * from './cart-product/cart-product.component';