+++ src/app/core/models/AppInfo.model.ts
@@ -1,0 +1,4 @@
+export interface AppInfo {
+  App: string;
+  Ver: string;
+}