+++ src/app/core/models/UserConfig.ts
@@ -1,0 +1,6 @@
+export interface UserConfig {
+  id?: string;
+  login?: string;
+  email?: string;
+  age?: number;
+}