+++ src/app/core/models/index.ts
@@ -1,0 +1,2 @@
+export * from './AppInfo.model';
+export * from './UserConfig';