+++ src/app/core/services/constants.service.ts
@@ -1,0 +1,6 @@
+import { AppInfo } from '../models';
+
+export const ConstantService: AppInfo = {
+  App: 'Shop',
+  Ver: '1.0',
+};