+++ src/app/layout/layout.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { AboutComponent } from './components';
+
+import { SharedModule } from '../shared/shared.module';
+
+@NgModule({
+  declarations: [AboutComponent],
+  imports: [
+    CommonModule,
+    SharedModule,
+  ],
+  exports: [AboutComponent],
+})
+export class LayoutModule { }