+++ src/app/products/components/index.ts
@@ -1,2 +1,2 @@
-export * from './product-list';
-export * from './product';
+export * from './product-list/product-list.component';
+export * from './product/product.component';